split media_player into package
