+++ custom_components/magic_areas/binary_sensor/ble_tracker.py
@@ -3,18 +3,15 @@
 from datetime import UTC, datetime
 import logging
 
-from homeassistant.helpers.event import async_track_state_change_event
 from homeassistant.components.binary_sensor import (
     DOMAIN as BINARY_SENSOR_DOMAIN,
     BinarySensorDeviceClass,
     BinarySensorEntity,
 )
 from homeassistant.const import ATTR_ENTITY_ID, STATE_ON
-from homeassistant.core import (
-    Event,
-    EventStateChangedData,
-    callback,
-)
+from homeassistant.core import Event, EventStateChangedData, callback
+from homeassistant.helpers.event import async_track_state_change_event
+
 from custom_components.magic_areas.base.entities import MagicEntity
 from custom_components.magic_areas.base.magic import MagicArea
 from custom_components.magic_areas.const import (
